grouped_mm autocast fix: custom ops bypass amp casting - bf16 activations x fp32 master weights crashed tl.dot on the 8x launch; cast to autocast dtype in wrapper (autograd-tracked, fp32 grads preserved) + regression test

+++ core/fourstroke.py
@@ -280,7 +280,8 @@ class AttnBackend(nn.Module):
             cg = cflat.index_select(0, lin)               # (n, d)
             qg = grouped_mm(self.ln_q(cg), self.w_q.weight, offs, tg, tr)
             q = self._heads(torch.zeros_like(cflat)
-                            .index_copy(0, lin, qg).view(B, K, T, d))
+                            .index_copy(0, lin, qg.to(cflat.dtype))
+                            .view(B, K, T, d))
         elif disp is None:
             q = self._heads(self.w_q(self.ln_q(c_prev)))
         else:
